fix octane mc configuration
